Resolve "Authentifizierung und Authorisierung muss einbauen"

+++ backend/src/routes/animals.ts
@@ -1,54 +1,53 @@
-import express, { Request } from "express";
-import { AnimalracesType, AnimalsCreateSchema, AnimalUpdateSchema, Animal_has_RacesCreateSchema, Animal_has_RacesType, AddRacesToAnimalSchema } from "vetlib-shared/schemas/ZodSchemas";
+import express, { Request, Response, NextFunction } from "express";
+import { AnimalracesType, AnimalsCreateSchema, AnimalUpdateSchema, Animal_has_RacesCreateSchema, Animal_has_RacesType, AddRacesToAnimalSchema, AnimalsCreateType, PostgresIdSchema } from "vetlib-shared/schemas/ZodSchemas";
 import { animalService } from "../service/animalService";
 import { personService } from "../service/personService";
 import { animalRaceService } from "../service/animalRaceService";
 import { animalHasRacesService } from "../service/animalHasRacesService";
 import multer from "multer";
 import { optionalAuthentication, requiresAuthentication } from "./authentication";
+import { ConstraintError } from "../exceptions/errors/ContraintError";
+import { AuthorizationError } from "../exceptions/errors/AuthorizationError";
 export const animalsRouter = express.Router();
 
 const storage = multer.diskStorage({
-  destination: (req, file, cb) => {
-    cb(null, 'public/uploads/animals'); // make sure this folder exists
-  },
-  filename: (req, file, cb) => {
-    const uniqueSuffix = Date.now() + '-' + Math.round(Math.random() * 1E9);
-    cb(null, file.fieldname + '-' + uniqueSuffix + '.' + file.originalname.split('.').pop());
-  }
+    destination: (req, file, cb) => {
+        cb(null, 'public/uploads/animals'); // make sure this folder exists
+    },
+    filename: (req, file, cb) => {
+        const uniqueSuffix = Date.now() + '-' + Math.round(Math.random() * 1E9);
+        cb(null, file.fieldname + '-' + uniqueSuffix + '.' + file.originalname.split('.').pop());
+    }
 });
 
 const upload = multer({
-  storage: storage,
-  limits: { fileSize: 2 * 1024 * 1024 }, // 2MB limit
-  fileFilter: (req, file, cb) => {
-    if (!file.mimetype.startsWith('image/')) {
-      return cb(new Error('Only images are allowed'));
+    storage: storage,
+    limits: { fileSize: 2 * 1024 * 1024 }, // 2MB limit
+    fileFilter: (req, file, cb) => {
+        if (!file.mimetype.startsWith('image/')) {
+            return cb(new Error('Only images are allowed'));
+        }
+        cb(null, true);
     }
-    cb(null, true);
-  }
 });
+
+async function ensureUserCanAccessAnimal(personId: number, animalId: number) {
+    const hasAccess = await animalService.canPersonAccessAnimal(personId, animalId);
+    if (!hasAccess) {
+        throw new AuthorizationError("No access");
+    }
+}
 
 animalsRouter.post("/",
     requiresAuthentication,
     async (req, res) => {
-        const parseResult = AnimalsCreateSchema.safeParse(req.body);
-        if (!parseResult.success) {
-            res.status(400).send(parseResult.error);
-            return;
-        }
-
-        // TODO: After implementing validation get the personId or the
-        // identifier for the person from the authentification token
-        const personId = parseInt(req.userId!);
-
-        const creationData = parseResult.data;
+        const validatedBody = AnimalsCreateSchema.parse(req.body);
 
         // Create new animal
-        const animal = await animalService.create(creationData);
+        const animal = await animalService.create(validatedBody);
 
         // Connect animal to requestor
-        await personService.connectAnimal(personId, animal.id);
+        await personService.connectAnimal(req.userId!, animal.id);
 
         res.send(animal);
     }
@@ -57,19 +56,19 @@ animalsRouter.post("/",
 animalsRouter.put('/:animalId',
     requiresAuthentication,
     async (req, res) => {
-        const parseResult = AnimalUpdateSchema.safeParse(req.body);
-        if (!parseResult.success) {
-            res.status(400).send(parseResult.error);
-            return;
+        const validatedBody = AnimalUpdateSchema.parse(req.body);
+        const animalId = PostgresIdSchema.parse(parseInt(req.params.animalId));
+
+        if (animalId !== validatedBody.id) {
+            throw new ConstraintError("Mismatch", [
+                { path: 'query.id', value: animalId },
+                { path: 'body.id', value: validatedBody.id }
+            ]);
         }
 
-        const animalId = parseInt(req.params.animalId);
-        if (animalId !== parseResult.data.id) {
-            res.status(400).send(`Mismatch in param id '${animalId}' and provided body id '${parseResult.data.id}'.`);
-            return;
-        }
+        ensureUserCanAccessAnimal(req.userId!, validatedBody.id);
 
-        const animal = await animalService.update(parseResult.data);
+        const animal = await animalService.update(validatedBody);
         return res.send(animal);
     }
 );
@@ -77,7 +76,10 @@ animalsRouter.put('/:animalId',
 animalsRouter.get('/:animalId/picture',
     requiresAuthentication,
     async (req, res) => {
-        const animalId = parseInt(req.params.animalId);
+        const animalId = PostgresIdSchema.parse(parseInt(req.params.animalId));
+
+        ensureUserCanAccessAnimal(req.userId!, animalId);
+
         const filepath = await animalService.getPicturePath(animalId);
         res.sendFile(filepath);
     }
@@ -87,7 +89,10 @@ animalsRouter.post('/:animalId/picture',
     requiresAuthentication,
     upload.single('picture'),
     async (req, res) => {
-        const animalId = parseInt(req.params.animalId);
+        const animalId = PostgresIdSchema.parse(parseInt(req.params.animalId));
+
+        ensureUserCanAccessAnimal(req.userId!, animalId);
+
         await animalService.savePicture(animalId, req.file?.path ?? null);
     }
 )
@@ -95,20 +100,22 @@ animalsRouter.post('/:animalId/picture',
 animalsRouter.delete('/:animalId',
     requiresAuthentication,
     async (req, res) => {
-        const animalId = parseInt(req.params.animalId);
-        try {
-            await animalService.delete(animalId);
-            res.sendStatus(204);
-        } catch (error) {
-            res.sendStatus(500);
-        }
+        const animalId = PostgresIdSchema.parse(parseInt(req.params.animalId));
+
+        ensureUserCanAccessAnimal(req.userId!, animalId);
+
+        await animalService.delete(animalId);
+        res.sendStatus(204);
     }
 )
 
 animalsRouter.get('/:animalId/races',
     requiresAuthentication,
     async (req, res) => {
-        const animalId = parseInt(req.params.animalId);
+        const animalId = PostgresIdSchema.parse(parseInt(req.params.animalId));
+
+        ensureUserCanAccessAnimal(req.userId!, animalId);
+
         const animalRaces: AnimalracesType[] = await animalRaceService.getAnimalRaces(animalId);
         return res.send(animalRaces);
     }
@@ -117,53 +124,50 @@ animalsRouter.get('/:animalId/races',
 animalsRouter.post('/:animalId/races',
     requiresAuthentication,
     async (req, res) => {
-        const animalId = parseInt(req.params.animalId);
-        const animalRaceData = AddRacesToAnimalSchema.safeParse(req.body);
-        if (!animalRaceData.success) {
-            res.status(400).send(animalRaceData.error);
+        const animalId = PostgresIdSchema.parse(parseInt(req.params.animalId));
+        const validatedBody = AddRacesToAnimalSchema.parse(req.body);
+
+        if (animalId !== validatedBody.animalid) {
+            res.status(400).send(`Mismatch in param id '${animalId}' and provided body id '${validatedBody.animalid}'.`);
             return;
         }
-        if (animalId !== animalRaceData.data.animalid) {
-            res.status(400).send(`Mismatch in param id '${animalId}' and provided body id '${animalRaceData.data.animalid}'.`);
-            return;
-        }
-        try {
-            const animalRace = await animalHasRacesService.create({
-                animalid: animalRaceData.data.animalid,
-                animalraceids: animalRaceData.data.animalraceids
-            });
-            res.status(201).send(animalRace);
-        } catch (error) {
-            res.sendStatus(500);
-        }
+
+        ensureUserCanAccessAnimal(req.userId!, animalId);
+        
+        const animalRace = await animalHasRacesService.create({
+            animalid: validatedBody.animalid,
+            animalraceids: validatedBody.animalraceids
+        });
+
+        res.status(201).send(animalRace);
     }
 )
 
 animalsRouter.delete('/:animalId/races',
     requiresAuthentication,
     async (req, res) => {
-        const animalId = parseInt(req.params.animalId);
-        try {
-            await animalHasRacesService.deleteAllRacesFromAnimal(animalId);
-            res.sendStatus(204);
-        } catch (error) {
-            res.sendStatus(500);
-        }
+        const animalId = PostgresIdSchema.parse(parseInt(req.params.animalId));
+
+        ensureUserCanAccessAnimal(req.userId!, animalId);
+
+        await animalHasRacesService.deleteAllRacesFromAnimal(animalId);
+        res.sendStatus(204);
     }
-)
+);
 
 animalsRouter.delete('/:animalId/races/:raceId',
     requiresAuthentication,
     async (req, res) => {
-        try {
-            await animalHasRacesService.delete({
-                fk_animalid: parseInt(req.params.animalId),
-                fk_animalraceid: parseInt(req.params.raceId)
-            });
-            res.sendStatus(204);
-        } catch (error) {
-            res.sendStatus(500);
-        }
-    }
-)
+        const animalId = PostgresIdSchema.parse(parseInt(req.params.animalId));
+        const raceId = PostgresIdSchema.parse(parseInt(req.params.raceId));
 
+        ensureUserCanAccessAnimal(req.userId!, animalId);
+
+        await animalHasRacesService.delete({
+            fk_animalid: animalId,
+            fk_animalraceid: raceId
+        });
+
+        res.sendStatus(204);
+    }
+);
